Extract recordings table formatters
