+++ Source/Parsek/UI/RecordingsTableUI.cs
@@ -3688,24 +3688,13 @@ namespace Parsek
             => ParsekTimeFormat.FormatDuration(seconds);
 
         internal static string FormatAltitude(double meters)
-        {
-            if (meters < 1000) return $"{(int)meters}m";
-            if (meters < 1000000) return (meters / 1000).ToString("F1", System.Globalization.CultureInfo.InvariantCulture) + "km";
-            return (meters / 1000000).ToString("F1", System.Globalization.CultureInfo.InvariantCulture) + "Mm";
-        }
+            => RecordingsTableFormatters.FormatAltitude(meters);
 
         internal static string FormatSpeed(double mps)
-        {
-            if (mps < 1000) return $"{(int)mps}m/s";
-            return (mps / 1000).ToString("F1", System.Globalization.CultureInfo.InvariantCulture) + "km/s";
-        }
+            => RecordingsTableFormatters.FormatSpeed(mps);
 
         internal static string FormatDistance(double meters)
-        {
-            if (meters < 1000) return $"{(int)meters}m";
-            if (meters < 1000000) return (meters / 1000).ToString("F1", System.Globalization.CultureInfo.InvariantCulture) + "km";
-            return (meters / 1000000).ToString("F1", System.Globalization.CultureInfo.InvariantCulture) + "Mm";
-        }
+            => RecordingsTableFormatters.FormatDistance(meters);
 
         /// <summary>
         /// Resolves the parent vessel name for EVA recordings by looking up ParentRecordingId.
@@ -3731,24 +3720,7 @@ namespace Parsek
         /// Priority: launch site > EVA from vessel > situation + biome + body > biome + body > body.
         /// </summary>
         internal static string FormatStartPosition(Recording rec, string parentVesselName = null)
-        {
-            // EVA: show source vessel
-            if (!string.IsNullOrEmpty(rec.EvaCrewName))
-            {
-                if (!string.IsNullOrEmpty(parentVesselName))
-                    return "EVA from " + parentVesselName;
-                return FormatSituationLocation(rec.StartSituation, rec.StartBiome, rec.StartBodyName, "EVA");
-            }
-
-            // Launch from a site
-            if (!string.IsNullOrEmpty(rec.LaunchSiteName))
-                return !string.IsNullOrEmpty(rec.StartBodyName)
-                    ? rec.LaunchSiteName + ", " + rec.StartBodyName
-                    : rec.LaunchSiteName;
-
-            // General: use situation + biome + body
-            return FormatSituationLocation(rec.StartSituation, rec.StartBiome, rec.StartBodyName, null);
-        }
+            => RecordingsTableFormatters.FormatStartPosition(rec, parentVesselName);
 
         /// <summary>
         /// Formats the recording end position for the expanded stats column.
@@ -3757,59 +3729,7 @@ namespace Parsek
         /// Body fallback priority for mid-segments: SegmentBodyName → last point body → StartBodyName.
         /// </summary>
         internal static string FormatEndPosition(Recording rec, string parentVesselName = null)
-        {
-            if (!rec.TerminalStateValue.HasValue)
-            {
-                // No terminal state (chain mid-segment or interior tree recording).
-                // Fallback: SegmentBodyName → last trajectory point body → StartBodyName.
-                string segBody = rec.SegmentBodyName;
-                if (string.IsNullOrEmpty(segBody) && rec.Points != null && rec.Points.Count > 0)
-                    segBody = rec.Points[rec.Points.Count - 1].bodyName;
-                if (string.IsNullOrEmpty(segBody))
-                    segBody = rec.StartBodyName;
-
-                string segmentLabel = RecordingStore.GetSegmentPhaseLabel(rec);
-                if (!string.IsNullOrEmpty(segmentLabel))
-                    return segmentLabel;
-                if (!string.IsNullOrEmpty(segBody))
-                    return segBody;
-                return "-";
-            }
-
-            string body = rec.TerminalOrbitBody;
-            if (string.IsNullOrEmpty(body) && !string.IsNullOrEmpty(rec.StartBodyName))
-                body = rec.StartBodyName;
-
-            switch (rec.TerminalStateValue.Value)
-            {
-                case TerminalState.Orbiting:
-                    return !string.IsNullOrEmpty(body) ? "Orbiting " + body : "Orbiting";
-                case TerminalState.Docked:
-                    return !string.IsNullOrEmpty(body) ? "Docked, " + body : "Docked";
-
-                case TerminalState.Landed:
-                case TerminalState.Splashed:
-                    if (!string.IsNullOrEmpty(rec.EndBiome) && !string.IsNullOrEmpty(body))
-                        return rec.EndBiome + ", " + body;
-                    if (!string.IsNullOrEmpty(body))
-                        return body;
-                    return rec.TerminalStateValue.Value.ToString();
-
-                case TerminalState.Destroyed:
-                    return !string.IsNullOrEmpty(body) ? "Destroyed, " + body : "Destroyed";
-                case TerminalState.Recovered:
-                    return !string.IsNullOrEmpty(body) ? "Recovered, " + body : "Recovered";
-                case TerminalState.SubOrbital:
-                    return !string.IsNullOrEmpty(body) ? "SubOrbital, " + body : "SubOrbital";
-                case TerminalState.Boarded:
-                    return !string.IsNullOrEmpty(parentVesselName)
-                        ? "Boarded " + parentVesselName
-                        : "Boarded";
-
-                default:
-                    return "-";
-            }
-        }
+            => RecordingsTableFormatters.FormatEndPosition(rec, parentVesselName);
 
         /// <summary>
         /// Formats a resource manifest for tooltip display.
@@ -3820,63 +3740,7 @@ namespace Parsek
         internal static string FormatResourceManifest(
             Dictionary<string, ResourceAmount> start,
             Dictionary<string, ResourceAmount> end)
-        {
-            if (start == null && end == null)
-                return null;
-
-            // Merge keys from both dicts
-            var keys = new SortedSet<string>();
-            if (start != null)
-                foreach (var k in start.Keys) keys.Add(k);
-            if (end != null)
-                foreach (var k in end.Keys) keys.Add(k);
-
-            if (keys.Count == 0)
-                return null;
-
-            bool hasEnd = end != null;
-            var lines = new List<string>();
-            lines.Add(hasEnd ? "Resources:" : "Resources at start:");
-
-            foreach (var key in keys)
-            {
-                if (hasEnd)
-                {
-                    double startAmt = 0;
-                    double endAmt = 0;
-                    if (start != null && start.TryGetValue(key, out var startRa))
-                        startAmt = startRa.amount;
-                    if (end.TryGetValue(key, out var endRa))
-                        endAmt = endRa.amount;
-
-                    double delta = endAmt - startAmt;
-                    string sign = delta >= 0 ? "+" : "";
-                    lines.Add(string.Format(CultureInfo.InvariantCulture,
-                        "  {0}: {1:F1} \u2192 {2:F1} ({3}{4:F1})",
-                        key,
-                        startAmt,
-                        endAmt,
-                        sign,
-                        delta));
-                }
-                else
-                {
-                    // Start only — show amount / maxAmount
-                    double amt = 0;
-                    double max = 0;
-                    if (start.TryGetValue(key, out var ra))
-                    {
-                        amt = ra.amount;
-                        max = ra.maxAmount;
-                    }
-                    lines.Add(string.Format(CultureInfo.InvariantCulture,
-                        "  {0}: {1:F1} / {2:F1}",
-                        key, amt, max));
-                }
-            }
-
-            return string.Join("\n", lines);
-        }
+            => RecordingsTableFormatters.FormatResourceManifest(start, end);
 
         /// <summary>
         /// Formats an inventory manifest for tooltip display.
@@ -3887,59 +3751,7 @@ namespace Parsek
         internal static string FormatInventoryManifest(
             Dictionary<string, InventoryItem> start,
             Dictionary<string, InventoryItem> end)
-        {
-            if (start == null && end == null)
-                return null;
-
-            // Merge keys from both dicts
-            var keys = new SortedSet<string>();
-            if (start != null)
-                foreach (var k in start.Keys) keys.Add(k);
-            if (end != null)
-                foreach (var k in end.Keys) keys.Add(k);
-
-            if (keys.Count == 0)
-                return null;
-
-            bool hasEnd = end != null;
-            var lines = new List<string>();
-            lines.Add(hasEnd ? "Inventory:" : "Inventory at start:");
-
-            foreach (var key in keys)
-            {
-                if (hasEnd)
-                {
-                    int startCount = 0;
-                    int endCount = 0;
-                    if (start != null && start.TryGetValue(key, out var startItem))
-                        startCount = startItem.count;
-                    if (end.TryGetValue(key, out var endItem))
-                        endCount = endItem.count;
-
-                    int delta = endCount - startCount;
-                    string sign = delta >= 0 ? "+" : "";
-                    lines.Add(string.Format(CultureInfo.InvariantCulture,
-                        "  {0}: {1} \u2192 {2} ({3}{4})",
-                        key,
-                        startCount,
-                        endCount,
-                        sign,
-                        delta));
-                }
-                else
-                {
-                    // Start only — show count
-                    int count = 0;
-                    if (start.TryGetValue(key, out var item))
-                        count = item.count;
-                    lines.Add(string.Format(CultureInfo.InvariantCulture,
-                        "  {0}: {1}",
-                        key, count));
-                }
-            }
-
-            return string.Join("\n", lines);
-        }
+            => RecordingsTableFormatters.FormatInventoryManifest(start, end);
 
         /// <summary>
         /// Formats a crew manifest for tooltip display.
@@ -3950,85 +3762,7 @@ namespace Parsek
         internal static string FormatCrewManifest(
             Dictionary<string, int> start,
             Dictionary<string, int> end)
-        {
-            if (start == null && end == null)
-                return null;
-
-            // Merge keys from both dicts
-            var keys = new SortedSet<string>();
-            if (start != null)
-                foreach (var k in start.Keys) keys.Add(k);
-            if (end != null)
-                foreach (var k in end.Keys) keys.Add(k);
-
-            if (keys.Count == 0)
-                return null;
-
-            bool hasEnd = end != null;
-            var lines = new List<string>();
-            lines.Add(hasEnd ? "Crew:" : "Crew at start:");
-
-            foreach (var key in keys)
-            {
-                if (hasEnd)
-                {
-                    int startCount = 0;
-                    int endCount = 0;
-                    if (start != null && start.TryGetValue(key, out var sc))
-                        startCount = sc;
-                    if (end.TryGetValue(key, out var ec))
-                        endCount = ec;
-
-                    int delta = endCount - startCount;
-                    string sign = delta >= 0 ? "+" : "";
-                    lines.Add(string.Format(CultureInfo.InvariantCulture,
-                        "  {0}: {1} \u2192 {2} ({3}{4})",
-                        key,
-                        startCount,
-                        endCount,
-                        sign,
-                        delta));
-                }
-                else
-                {
-                    // Start only — show count
-                    int count = 0;
-                    if (start.TryGetValue(key, out var sc))
-                        count = sc;
-                    lines.Add(string.Format(CultureInfo.InvariantCulture,
-                        "  {0}: {1}",
-                        key, count));
-                }
-            }
-
-            return string.Join("\n", lines);
-        }
-
-        /// <summary>
-        /// Formats situation + biome + body into a compact location string.
-        /// "Flying, Shores, Kerbin" or "Orbiting, Kerbin" or "Kerbin" etc.
-        /// </summary>
-        private static string FormatSituationLocation(string situation, string biome, string body, string prefix)
-        {
-            // Build: "{prefix/situation}, {biome}, {body}" with missing parts omitted
-            bool hasSit = !string.IsNullOrEmpty(situation);
-            bool hasBiome = !string.IsNullOrEmpty(biome);
-            bool hasBody = !string.IsNullOrEmpty(body);
-
-            string label = prefix ?? (hasSit ? situation : null);
-
-            if (label != null && hasBiome && hasBody)
-                return label + ", " + biome + ", " + body;
-            if (label != null && hasBody)
-                return label + ", " + body;
-            if (hasBiome && hasBody)
-                return biome + ", " + body;
-            if (hasBody)
-                return body;
-            if (label != null)
-                return label;
-            return "-";
-        }
+            => RecordingsTableFormatters.FormatCrewManifest(start, end);
 
         /// <summary>
         /// Returns the earliest StartUT among the given descendant recording indices.
@@ -4921,7 +4655,8 @@ namespace Parsek
 
         /// <summary>
         /// Builds the group tree data structures used to render the recordings tree.
-        /// Pure data-computation -- no IMGUI calls.
+        /// No IMGUI calls, but not pure: normalizes permanent root groups and
+        /// falls back to the active scenario supersede list when none is passed.
         /// </summary>
         internal static void BuildGroupTreeData(
             IReadOnlyList<Recording> committed, int[] sortedIndices,
